backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
opencv-python==4.9.0.80
deepface==0.0.88
torch==2.1.2
ultralytics==8.1.0
numpy==1.26.3
Pillow==10.2.0
pydantic==2.5.3
python-dotenv==1.0.0
tensorflow==2.13.0
scikit-learn==1.3.0
joblib==1.3.2